refactor api

+++ API/src/main/java/com/packshop/api/services/category/CategoryService.java
@@ -1,13 +1,15 @@
 package com.packshop.api.services.category;
 
-import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.packshop.api.entities.category.Category;
 import com.packshop.api.repositories.category.CategoryRepository;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CategoryService {
@@ -19,15 +21,31 @@ public class CategoryService {
         return categoryRepository.findAll();
     }
 
-    public Optional<Category> getCategoryById(Long id) {
-        return categoryRepository.findById(id);
+    public ResponseEntity<Category> getCategoryById(Long id) {
+        Optional<Category> category = categoryRepository.findById(id);
+        return category.map(ResponseEntity::ok)
+                .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-    public Category saveCategory(Category category) {
-        return categoryRepository.save(category);
+    public ResponseEntity<Category> createCategory(Category category) {
+        Category savedCategory = categoryRepository.save(category);
+        return new ResponseEntity<>(savedCategory, HttpStatus.CREATED);
     }
 
-    public void deleteCategory(Long id) {
+    public ResponseEntity<Category> updateCategory(Long id, Category category) {
+        if (!categoryRepository.existsById(id)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        category.setId(id);
+        Category updatedCategory = categoryRepository.save(category);
+        return ResponseEntity.ok(updatedCategory);
+    }
+
+    public ResponseEntity<Void> deleteCategory(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         categoryRepository.deleteById(id);
+        return ResponseEntity.noContent().build();
     }
 }
